% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createcb.R
\name{createcb}
\alias{createcb}
\title{Create a codebook}
\usage{
createcb(data, headers = TRUE, cbname = "codebook")
}
\arguments{
\item{data}{Dataframe to create codebook}

\item{headers}{Default = TRUE; Assumes variable names are treated as headers and row 1 contains variable descriptions}

\item{cbname}{Default = "codebook"; Name of codebook to create}
}
\value{
Formatted dataframe with 1 header row and codebook with 5 columns (Variable, Alias, Description, Scoring, Citation)
}
\description{
Create a codebook
}
