% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportclean.R
\name{export}
\alias{export}
\title{Export clean data and codebook}
\usage{
export(
  data,
  name,
  filetype = "csv",
  version = "all",
  idcols = c("IPAddress", "LocationLatitude", "LocationLongitude", "ExternalReference",
    "RecipientFirstName", "RecipientLastName", "RecipientEmail"),
  usecb = TRUE,
  cbname = "codebook"
)
}
\arguments{
\item{data}{Dataframe to create codebook}

\item{name}{Path name for exported file(s)}

\item{filetype}{Default = "csv"; "sav" exports .sav file (for use with SPSS), "all" exports .sav and .csv files}

\item{version}{Default = "all"; "annotated" exports de-identified data with no exclusions, "clean" exports de-identified data with exclusions, "as-is" exports data in current format}

\item{idcols}{List of identifying columns to be excluded when de-identifying data; Default = c("IPAddress","LocationLatitude","LocationLongitude","ExternalReference","RecipientFirstName","RecipientLastName","RecipientEmail")}

\item{usecb}{Default = TRUE; use "Description" column in codebook as labels in .sav exports}

\item{cbname}{Default = "codebook"; name of codebook to use}
}
\description{
Export clean data and codebook
}
