% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{extreme}
\alias{extreme}
\title{Check for outliers}
\usage{
extreme(data, var, threshold, exclude = FALSE)
}
\arguments{
\item{data}{Dataframe containing column to check for outliers}

\item{var}{Name of column to check for outliers (must be enclosed in double quotes)}

\item{threshold}{Number in standard deviation units to act as upper and lower bound for outliers}

\item{exclude}{Default = FALSE = returns list of outliers; TRUE= returns data excluding outliers}
}
\description{
Check for outliers
}
