% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_IP.R
\name{ip_check}
\alias{ip_check}
\title{Check IP addresses using IPhub API}
\usage{
ip_check(data, ip, api_key, updatecb = TRUE, cbname = "codebook")
}
\arguments{
\item{data}{Dataframe containing IP addresses you want to check}

\item{ip}{Name of column in Dataframe containing IP addresses}

\item{api_key}{Your API key obtained from IPhub}

\item{updatecb}{Default =TRUE; Updates existing codebook with IPflag and IPcountry variables}

\item{cbname}{Default = "codebook"; Name of codebook to update}
}
\value{
Dataframe with 2 new columns (IPflag and IPcountry). IPflag contains values 0,1, or 2. IPcountry contains country code of IP address.
}
\description{
Check IP addresses using IPhub API
}
