% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_IP.R
\name{ip_filter}
\alias{ip_filter}
\title{Filter data based on IPcountry and IPflag}
\usage{
ip_filter(
  data,
  cval,
  fval,
  country = "IPcountry",
  flag = "IPflag",
  exclude = FALSE,
  intersect = FALSE
)
}
\arguments{
\item{data}{Dataframe to filter}

\item{cval}{country code value in 'IPcountry' column you would like to filter by}

\item{fval}{value in 'IPflag' column you would like to filter by}

\item{country}{name of variable containing country codes (must be enclosed in double quotes); defaults to "IPcountry"}

\item{flag}{name of variable containing flags (must be enclosed in double quotes); defaults to "IPflag"}

\item{exclude}{Default = FALSE = returns list of cases with specified country code and/or flag values; TRUE = returns data excluding cases with specified country code or flag values}

\item{intersect}{Default = FALSE = identifies cases not specified country code OR not specified flag; TRUE = identifies cases not specified country code AND not specified flag}
}
\value{
Filtered data excluding cases with specified IPcountry or IPflag values
}
\description{
Filter data based on IPcountry and IPflag
}
