% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Rename variables}
\usage{
rename(data, old, new, updatecb = TRUE, cbname = "codebook")
}
\arguments{
\item{data}{Dataframe}

\item{old}{Variable to rename}

\item{new}{New name for variable}

\item{updatecb}{Default=TRUE}

\item{cbname}{Default = "codebook"; Name of codebook to update}
}
\value{
Dataframe with renamed variable
}
\description{
Rename variables
}
