% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescore.R
\name{rescore}
\alias{rescore}
\title{Recode and reverse score items}
\usage{
rescore(data, vars, old, new = NULL, updatecb = TRUE, cbname = "codebook")
}
\arguments{
\item{data}{Dataframe to create codebook}

\item{vars}{Name of item to be reverse scored}

\item{old}{List of values present in variable}

\item{new}{Default = NULL}

\item{updatecb}{Default=TRUE}

\item{cbname}{Default = "codebook"; Name of codebook to update}
}
\value{
Dataframe with recoded variables
}
\description{
Recode and reverse score items
}
