% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_dup.R
\name{fraud_check}
\alias{fraud_check}
\title{Screen for bots using Qualtrics bot detection parameters}
\usage{
fraud_check(
  data,
  recaptcha = "Q_RecaptchaScore",
  relID = "Q_RelevantIDFraudScore",
  rclimit = 0.5,
  relIDlimit = 30,
  exclude = FALSE,
  intersect = FALSE
)
}
\arguments{
\item{data}{Dataframe containing bot detection variables}

\item{recaptcha}{Name of reCAPTCHA variable; defaults to Q_RecaptchaScore, assuming Qualtrics naming scheme}

\item{relID}{Name of RelevantID fraud variable; defaults to Q_RelevantIDFraudScore, assuming Qualtrics naming scheme}

\item{rclimit}{reCAPTCHA yhreshold to use for flagging responses; defaults to 0.5; cases scoring less than this value will be flagged}

\item{relIDlimit}{RelevantID threshold to use for flagging responses; defaults to 30; cases scoring greater than or equal to this value will be flagged}

\item{exclude}{Default = FALSE = identifies cases that are likely bots based on recaptcha/relID thresholds set; TRUE = returns data excluding likely bots}

\item{intersect}{Default = FALSE = identifies cases based on recaptcha OR relID; TRUE = identifies cases based on recaptcha AND relID}
}
\description{
Screen for bots using Qualtrics bot detection parameters
}
