% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbe.R
\name{tbe}
\alias{tbe}
\title{Add exclusion filter variable and annotate reason for exclusion (tbe = to be excluded)}
\usage{
tbe(
  data,
  var1,
  val1,
  v1prime = FALSE,
  var2,
  val2,
  v2prime = FALSE,
  intersect = FALSE,
  updatecb = TRUE,
  cbname = "codebook"
)
}
\arguments{
\item{data}{Dataframe}

\item{var1}{Name of column to filter (must be enclosed in double quotes)}

\item{val1}{Value of var1 to filter}

\item{v1prime}{Default = FALSE = indicates you want to exclude cases where var1=val1; TRUE = indicates you want to exclude cases where var1 is NOT equal to val1}

\item{var2}{Name of 2nd column to filter (must be enclosed in double quotes)}

\item{val2}{Value of var2 to filter}

\item{v2prime}{Default = FALSE = indicates you want to exclude cases where var2=val2; TRUE = indicates you want to exclude cases where var2 is NOT equal to val2}

\item{intersect}{Default = FALSE = identifies cases meeting criteria for var1 OR var2; TRUE = identifies cases meeting criteria for var 1 AND var2}

\item{updatecb}{Default = TRUE = updates codebook with new filter variables}

\item{cbname}{Default = "codebook" = name of codebook to be updated}
}
\value{
Dataframe with exclusion columns
}
\description{
Add exclusion filter variable and annotate reason for exclusion (tbe = to be excluded)
}
